Fixing more warnings.

+++ integration/non_event_data/src/BasicGeometry.cpp
@@ -35,7 +35,7 @@ void Consumers::BasicGeometry::consume(std::vector<char> const& data)
     cudaCheck(cudaMalloc((void**) &p, data.size()));
     dev_geometry = gsl::span {p, static_cast<gsl::span<char>::index_type>(data.size())};
   }
-  else if (dev_geometry.size() != data.size()) {
+  else if ((size_t) dev_geometry.size() != data.size()) {
     throw StrException {string {"sizes don't match: "} + to_string(dev_geometry.size()) + " " + to_string(data.size())};
   }
   cudaCheck(cudaMemcpy(dev_geometry.data(), data.data(), data.size(), cudaMemcpyHostToDevice));
